+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ValidationResourcesSpec.scala
@@ -116,6 +116,22 @@ class ValidationResourcesSpec
     }
   }
 
+  it should "find errors in scenario id" in {
+    createAndValidateScenario(ProcessTestData.validProcessWithId(" ")) {
+      status shouldEqual StatusCodes.OK
+      val entity = entityAs[String]
+      entity should include("Scenario name cannot be blank")
+    }
+  }
+
+  it should "find errors in node id" in {
+    createAndValidateScenario(ProcessTestData.validProcessWithNodeId(" ")) {
+      status shouldEqual StatusCodes.OK
+      val entity = entityAs[String]
+      entity should include("Node name cannot be blank")
+    }
+  }
+
   it should "return fatal error for bad ids" in {
     val invalidCharacters = newDisplayableProcess(
       "p1",
@@ -129,7 +145,7 @@ class ValidationResourcesSpec
     createAndValidateScenario(invalidCharacters) {
       status shouldEqual StatusCodes.BadRequest
       val entity = entityAs[String]
-      entity should include("Node f1\"' contains invalid characters")
+      entity should include("Node name contains invalid characters")
     }
 
     val duplicateIds = newDisplayableProcess(